Issue#24- recipeview and sidebar cooperation
