+++ public/index.html
@@ -21,7 +21,7 @@
     -->
     <title>React App</title>
   </head>
-  <body>
+  <body style="height: 100vh; overflow-y: hidden;">
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
