+++ src/components/Sidebar.js
@@ -27,7 +27,7 @@ class Sidebar extends Component {
       </ListItem>
     ));
     return (
-      <GrommetSidebar size='small' separator='all'>
+      <GrommetSidebar fixed={true} size='small' separator='all'>
         <Title align='center'>Ingredients</Title>
         <Search
           inline={true} suggestions={library}
